trying to synchronize metadata

+++ hyracks-fullstack/hyracks/hyracks-control/hyracks-control-cc/src/main/java/org/apache/hyracks/control/cc/job/IJobManager.java
@@ -151,7 +151,7 @@ public interface IJobManager {
      */
     long getTotalRejectedJobs();
 
-    public default int getDefaultQueuePriority() {
+    public default long getDefaultQueuePriority() {
         return 1;
     }
 }
