+++ hyracks-fullstack/hyracks/hyracks-control/hyracks-control-cc/src/main/java/org/apache/hyracks/control/cc/job/WorkloadManager.java
@@ -111,7 +111,7 @@ public class WorkloadManager extends JobManager {
     }
 
     @Override
-    public int getDefaultQueuePriority() {
+    public long getDefaultQueuePriority() {
         return jobTypeManager.getDefaultPriority();
     }
