+++ hyracks-fullstack/hyracks/hyracks-control/hyracks-control-cc/src/main/java/org/apache/hyracks/control/cc/scheduler/IJobTypeManager.java
@@ -6,5 +6,5 @@ public interface IJobTypeManager {
 
     void setJobType(JobRun jobRun);
 
-    int getDefaultPriority();
+    long getDefaultPriority();
 }
